Key symbols for key events
Add numeric symbols and #defines for most common keys in keyboards.
Keysyms with values less than 128 are equal to ASCII values that
correspond to the character from the key.

There isn't an exhaustive list of possible keys because it would be
impossible to do in a way that is portable, internationalizable and
maintainable.

Key symbols are supposed to be used only for hard-coded key bindings and
default keys. For user-programmable keys, native scan codes should be
used. TODO: add a function to look up names of keys based on their scan
code (not keysym).

Use qsort and bsearch to look for keysyms in a look up table. The array
is not pre-sorted to make it easier to maintain.